MAGECLOUD-951: Create preDeploy process
