+++ src/Magento/MagentoCloud/Process/Deploy/InstallUpdate.php
@@ -7,7 +7,7 @@ namespace Magento\MagentoCloud\Process\Deploy;
 
 use Magento\MagentoCloud\Config\DeploymentConfig;
 use Magento\MagentoCloud\DB\Adapter;
-use Magento\MagentoCloud\Environment;
+use Magento\MagentoCloud\Config\Environment;
 use Magento\MagentoCloud\Filesystem\Driver\File;
 use Magento\MagentoCloud\Process\ProcessInterface;
 use Magento\MagentoCloud\Shell\ShellInterface;
